refactor(viewer): consolidate workers into analyze.js
