--- dial9-viewer/toolkit/accumulate_worker.js
@@ -1,1 +1,0 @@
-../skills/accumulate_worker.js
