--- dial9-viewer/toolkit/parse_worker.js
@@ -1,1 +1,0 @@
-../skills/parse_worker.js
